requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
jinja2==3.1.4
python-multipart==0.0.9

numpy==1.26.4
scipy==1.11.4
soundfile==0.12.1

opencv-python==4.10.0.84
mediapipe==0.10.14

tensorflow-macos==2.15.0
tensorflow-metal==1.1.0

vosk==0.3.44
pydantic==2.8.2
